LabelSpecifiersを、抽象基底クラスとそれを継承するMetadataLabelSpecifiersに分離した
